# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "GeocodeResolutionPayload": {
        "additionalProperties": false,
        "properties": {
          "query": {
            "minLength": 1,
            "title": "Query",
            "type": "string"
          }
        },
        "required": [
          "query"
        ],
        "title": "GeocodeResolutionPayload",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MapGroupMembershipResponse": {
        "properties": {
          "group_id": {
            "title": "Group Id",
            "type": "string"
          },
          "group_type": {
            "enum": [
              "general",
              "application_wave"
            ],
            "title": "Group Type",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "group_id",
          "name",
          "group_type"
        ],
        "title": "MapGroupMembershipResponse",
        "type": "object"
      },
...
      "MapProjectionFeatureResponse": {
        "properties": {
          "availability": {
            "enum": [
              "available",
              "unavailable",
              "uncertain",
              "unknown"
            ],
            "title": "Availability",
            "type": "string"
          },
          "company_id": {
            "title": "Company Id",
            "type": "string"
          },
          "company_name": {
            "title": "Company Name",
            "type": "string"
          },
          "feature_id": {
            "title": "Feature Id",
            "type": "string"
          },
          "groups": {
            "items": {
              "$ref": "#/components/schemas/MapGroupMembershipResponse"
            },
            "title": "Groups",
            "type": "array"
          },
          "latitude": {
            "title": "Latitude",
            "type": "number"
          },
          "location_label": {
            "title": "Location Label",
            "type": "string"
          },
          "longitude": {
            "title": "Longitude",
            "type": "number"
          },
          "opportunity_id": {
            "title": "Opportunity Id",
            "type": "string"
          },
          "precision": {
            "title": "Precision",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "tracking_status": {
            "enum": [
              "new",
              "to_review",
              "interesting",
              "shortlisted",
              "deferred",
              "excluded",
              "archived"
            ],
            "title": "Tracking Status",
            "type": "string"
          },
          "work_location_id": {
            "title": "Work Location Id",
            "type": "string"
          }
        },
        "required": [
          "feature_id",
          "work_location_id",
          "opportunity_id",
          "company_id",
          "title",
          "company_name",
          "location_label",
          "latitude",
          "longitude",
          "precision",
          "tracking_status",
          "availability",
          "groups"
        ],
        "title": "MapProjectionFeatureResponse",
        "type": "object"
      },
      "MapProjectionPayload": {
        "additionalProperties": false,
        "properties": {
          "opportunity_ids": {
            "items": {
              "type": "string"
            },
            "title": "Opportunity Ids",
            "type": "array"
          }
        },
        "title": "MapProjectionPayload",
        "type": "object"
      },
      "MapResolutionPayload": {
        "additionalProperties": false,
        "properties": {
          "latitude": {
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "title": "Longitude",
            "type": "number"
          },
          "resolved_query": {
            "minLength": 1,
            "title": "Resolved Query",
            "type": "string"
          }
        },
        "required": [
          "latitude",
          "longitude",
          "resolved_query"
        ],
        "title": "MapResolutionPayload",
        "type": "object"
      },
      "MapResolutionResponse": {
        "properties": {
          "latitude": {
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "title": "Longitude",
            "type": "number"
          },
          "provider_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Key"
          },
          "resolution_source": {
            "enum": [
              "manual",
              "geocoder"
            ],
            "title": "Resolution Source",
            "type": "string"
          },
          "resolved_at": {
            "title": "Resolved At",
            "type": "string"
          },
          "resolved_query": {
            "title": "Resolved Query",
            "type": "string"
          }
        },
        "required": [
          "latitude",
          "longitude",
          "resolution_source",
          "provider_key",
          "resolved_at",
          "resolved_query"
        ],
        "title": "MapResolutionResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Vocation",
    "version": "0.4.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/map/locations/{work_location_id}/geocode": {
      "post": {
        "operationId": "geocode_location_api_map_locations__work_location_id__geocode_post",
        "parameters": [
          {
            "in": "path",
            "name": "work_location_id",
            "required": true,
            "schema": {
              "title": "Work Location Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeocodeResolutionPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MapResolutionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Geocode Location",
        "tags": [
          "map"
        ]
      }
    },
    "/api/map/locations/{work_location_id}/resolution": {
      "delete": {
        "operationId": "delete_resolution_api_map_locations__work_location_id__resolution_delete",
        "parameters": [
          {
            "in": "path",
            "name": "work_location_id",
            "required": true,
            "schema": {
              "title": "Work Location Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Resolution",
        "tags": [
          "map"
        ]
      },
      "put": {
        "operationId": "set_resolution_api_map_locations__work_location_id__resolution_put",
        "parameters": [
          {
            "in": "path",
            "name": "work_location_id",
            "required": true,
            "schema": {
              "title": "Work Location Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MapResolutionPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MapResolutionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Resolution",
        "tags": [
          "map"
        ]
      }
    },
    "/api/map/projection": {
      "post": {
        "operationId": "map_projection_api_map_projection_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MapProjectionPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MapProjectionFeatureResponse"
                  },
                  "title": "Response Map Projection Api Map Projection Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Map Projection",
        "tags": [
          "map"
        ]
      }
    },
...
  }
}